--- script/_functions
@@ -1,10 +1,0 @@
-set -e
-
-function command_available() {
-  command -v "$@" >/dev/null 2>&1
-}
-
-function warning() {
-  [[ $# -eq 0 ]] && local message=$(cat -) || local message="$@"
-  echo "$(tput setaf 3)$message$(tput sgr0)"
-}
