--- script/cordova-exec
@@ -1,15 +1,0 @@
-#!/usr/bin/env bash
-# This script allows to execute cordova commands with the app prepared for building.
-# It generates the HTML/JS/CSS etc and then executes cordova with any parameters you
-# provide to this script.
-#
-# This script is for example used by `bin/build` and `bin/run-on-device`.
-source "$(dirname ${BASH_SOURCE[0]})/_functions"
-PLATFORM=$2
-
-# build all assets
-bin/maji build
-
-export PATH=$(yarn bin):$PATH
-$(dirname ${BASH_SOURCE[0]})/prepare-cordova-platform "${@:2}"
-cd cordova; cordova "$@"
