--- script/create-project
@@ -1,43 +1,0 @@
-#!/usr/bin/env bash
-shopt -s dotglob # include hidden files in globs
-set -e # fail fast
-
-APP_PACKAGE=$1
-APP_NAME=$(echo $1 | perl -pe 's/.*\.//')
-APP_PATH="$2"
-
-MAJI_ROOT=$(dirname "${BASH_SOURCE[0]}")/../
-CURRENT_MAJI_VERSION=$(node -p "require('$MAJI_ROOT/package.json').version")
-MAJI_VERSION=${MAJI_VERSION:-$CURRENT_MAJI_VERSION}
-
-TMP_DIR=$(mktemp -d 2>/dev/null || mktemp -d -t maji)
-
-echo "Creating project in $APP_PATH ..."
-
-pushd "$MAJI_ROOT" >/dev/null 2>&1
-  cp -R project_template/* "$TMP_DIR"
-  script/create-package-json.js "$TMP_DIR" "$APP_PACKAGE" "$MAJI_VERSION"
-
-  pushd "$TMP_DIR" >/dev/null 2>&1
-    rm -rf node_modules
-    grep -rwl . -e '##APP_NAME##' | xargs perl -pi -e "s|##APP_NAME##|$APP_NAME|"
-    grep -rwl . -e '##APP_PACKAGE##' | xargs perl -pi -e "s|##APP_PACKAGE##|$APP_PACKAGE|"
-    grep -rwl . -e '##MAJI_VERSION##' | xargs perl -pi -e "s|##MAJI_VERSION##|$MAJI_VERSION|"
-
-    # create needed symlinks, NPM removes them
-    (cd cordova/ && ln -s ../dist www)
-    (cd bin/ && ln -s ../node_modules/.bin/maji .)
-
-    # npm strips .gitignore files during pack
-    [ -f gitignore ] && (mv gitignore .gitignore || exit 0)
-    command -v git >/dev/null && git init >/dev/null || true
-  popd >/dev/null 2>&1
-popd >/dev/null 2>&1
-
-mkdir -p "$APP_PATH"
-
-mv "$TMP_DIR"/* "$APP_PATH"
-rm -r "$TMP_DIR"
-
-cd "$APP_PATH"
-bin/setup
