--- script/prepare-cordova-platform
@@ -1,21 +1,0 @@
-#!/usr/bin/env bash
-source "$(dirname ${BASH_SOURCE[0]})/_functions"
-PLATFORM=$1
-
-pushd cordova >/dev/null 2>&1
-  cordova prepare $PLATFORM
-
-  if [ -f plugins.txt ]; then
-    warning <<EOF
-cordova/plugins.txt file detected.
-
-Maji plugin management is removed.
-Please use Cordova plugin management by configuring your plugins in cordova/config.xml,
-and remove cordova/plugins.txt.
-
-Refer to the Cordova documentation for details:
-https://cordova.apache.org/docs/en/latest/platform_plugin_versioning_ref/
-EOF
-    exit 1
-  fi
-popd >/dev/null 2>&1
